#frontend: cool it - emit rays

+++ frontend/src/components/shared/atoms/game/GameObject.vue
@@ -30,9 +30,8 @@ import GameContainer from '@/components/shared/atoms/game/GameContainer.vue';
 import { ObjectSpace } from '@/types/enum/ObjectSpace';
 import { delay } from '@/utils/wait';
 import { GrayscaleFilter } from 'pixi-filters';
-import Polygon from 'polygon';
-import Vec2 from 'vec2';
 import { toRadians, toDegrees } from '@/utils/angle';
+import * as matterUtil from '@/utils/matter';
 
 @Options({
   components: {},
@@ -305,26 +304,14 @@ export default class GameObject extends Vue {
     height: number,
     shape: [number, number][]
   ): void {
-    const calculateCentroid = (shape: [number, number][]): [number, number] => {
-      /*const p = new Polygon(shape.map((point) => new Vec2(point[0], point[1])));
-      const centroid = p.center(); */
-      /*const centroid: [number, number] = [0, 0];
-      shape.forEach((item) => {
-        centroid[0] += item[0];
-        centroid[1] += item[1];
-      });
-      centroid[0] /= shape.length;
-      centroid[1] /= shape.length;
-      return centroid;*/
+    this.options.isStatic = this.isStatic;
+    /*const calculateCentroid = (shape: [number, number][]): [number, number] => {
       const path = shape.map((item) => {
         return { x: item[0], y: item[1] };
       });
       const relativeBody = Matter.Bodies.fromVertices(0, 0, [path]);
       const minX = Math.min(...relativeBody.vertices.map((item) => item.x));
       const minY = Math.min(...relativeBody.vertices.map((item) => item.y));
-      /*const maxX = Math.max(...relativeBody.vertices.map((item) => item.x));
-      const maxY = Math.max(...relativeBody.vertices.map((item) => item.y));
-      return [100 - maxX, 100 - maxY];*/
 
       const minShapeX = Math.min(...shape.map((item) => item[0]));
       const minShapeY = Math.min(...shape.map((item) => item[1]));
@@ -334,7 +321,6 @@ export default class GameObject extends Vue {
       return [50 - deltaX, 50 - deltaY];
     };
 
-    this.options.isStatic = this.isStatic;
     const path = shape.map((item) => {
       return { x: (item[0] / 100) * width, y: (item[1] / 100) * height };
     });
@@ -343,7 +329,15 @@ export default class GameObject extends Vue {
     const deltaX = (width * (50 - centroid[0])) / 100;
     const deltaY = (height * (50 - centroid[1])) / 100;
     Matter.Body.setCentre(this.body, { x: deltaX, y: deltaY }, true);
-    Matter.Body.setPosition(this.body, { x: x, y: y });
+    Matter.Body.setPosition(this.body, { x: x, y: y });*/
+    this.body = matterUtil.createPolygonBody(
+      this.options,
+      x,
+      y,
+      width,
+      height,
+      shape
+    );
     this.updatePivot();
     this.onRotationChanged();
     this.onScaleChanged();
@@ -523,11 +517,13 @@ export default class GameObject extends Vue {
       if (index > -1) this.gameContainer.detector.bodies.splice(index, 1);
     }
     setTimeout(() => {
-      const parent = this.container.parent;
-      if (parent) {
-        parent.removeChild(this.container);
+      if (this.container) {
+        const parent = this.container.parent;
+        if (parent) {
+          parent.removeChild(this.container);
+        }
+        this.container.destroy({ children: true });
       }
-      this.container.destroy({ children: true });
     }, 100);
   }
